Add custom exception handler and logger

+++ src/CozyFileService.Application/Features/ManageFiles/Commands/DeleteFile/DeleteFileCommandHandler.cs
@@ -1,5 +1,7 @@
 ﻿using AutoMapper;
 using CozyFileService.Application.Contracts.Persistence;
+using CozyFileService.Application.Exceptions;
+using CozyFileService.Domain.Entities;
 using MediatR;
 
 namespace CozyFileService.Application.Features.ManageFiles.Commands.DeleteFile
@@ -18,6 +20,11 @@ namespace CozyFileService.Application.Features.ManageFiles.Commands.DeleteFile
         public async Task Handle(DeleteFileCommand request, CancellationToken cancellationToken)
         {
             var fileToDelete = await _uploadedFileRepository.GetByIdAsync(request.Id);
+
+            if (fileToDelete == null)
+            {
+                throw new NotFoundException(nameof(UploadedFile), request.Id);
+            }
 
             await _uploadedFileRepository.DeleteAsync(fileToDelete);
         }
